fx(aux_split_text[_blob]): fixes error `arithmetic exception, numeric overflow, or string truncation`
```
arithmetic exception, numeric overflow, or string truncation
numeric value is out of range
At procedure 'AUX_SPLIT_TEXT_BLOB' line: 49, col: 13
```

+++ prc_aux_split_text.sql
@@ -6,7 +6,7 @@ create or alter procedure aux_split_text(
     , trim_part smallint = 1
 )
 returns(
-    idx smallint
+    idx bigint
     , part varchar(32000)
 )
 as
